started toggle remove comment but it's still broken, fixed the edit post page

+++ src/components/CommentTile/CommentTile.jsx
@@ -1,17 +1,26 @@
 import "./CommentTile.css"
 import { Link } from "react-router";
 import VoteController from "../VoteController/VoteController";
-import { deleteComment } from "../../utils/comments";
+import { deleteComment, restoreComment } from "../../utils/comments";
 import { useState, useContext } from "react";
 import { UserContext } from "../../contexts/UserContext";
 
 export default function CommentTile ({comment}){
     const [error, setError] = useState(null)
     const {user} = useContext(UserContext)
+    const [isDeleted, setIsDeleted] = useState(comment.is_deleted)
 
-    const removeComment = () => {
+    const toggleRemoveComment = async() => {
+
+
+        
+
         try {
-            deleteComment(comment.id)
+            if(isDeleted){
+                await deleteComment(comment.id)
+            } else {
+                await restoreComment(comment.id)
+            }
         } catch (error) {
             setError(error)
         }
@@ -20,13 +29,14 @@ export default function CommentTile ({comment}){
     return(
             <div className="commentCard">
                 <div className="commentInfo">
-                    {comment.is_deleted ? <div className='userInfo'>[Anonymous User]</div> : <div className="userInfo">{comment.commenter.username}</div>}
-                    {user.id === comment.commenter.id && <button className="deleteButton" onClick={removeComment}>delete </button>}
+                    {isDeleted ? <div className='userInfo'>[Anonymous User]</div> : <div className="userInfo">{comment.commenter.username}</div>}
+                    {user.id === comment.commenter.id && 
+                        <button className={isDeleted ? "restoreButton" : "deleteButton"} onClick={toggleRemoveComment}>{isDeleted ? "Restore" : "Delete"} </button>}
                     
                 </div>
                 
                 <div className="commentTitle">
-                    {comment.is_deleted ? <p className="commentMessage">[Comment Removed]</p> : <p className="commentMessage">{comment.body}</p>}
+                    {isDeleted ? <p className="commentMessage">[Comment Removed]</p> : <p className="commentMessage">{comment.body}</p>}
                 </div>
                 <div className="commentInteractions">
                     <VoteController
